W-14875628 (refactor): added declaration parsers for Async 2.3 and 2.4
