+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/document/AsyncApi20DocumentParser.scala
@@ -1,7 +1,9 @@
 package amf.apicontract.internal.spec.async.parser.document
 
 import amf.apicontract.internal.spec.async.parser.context.AsyncWebApiContext
+import amf.apicontract.internal.spec.async.parser.domain.declarations.AsyncDeclarationParser
 import amf.core.internal.parser.Root
 
-case class AsyncApi20DocumentParser(root: Root)(override implicit val ctx: AsyncWebApiContext)
-    extends AsyncApiDocumentParser(root)(ctx)
+case class AsyncApi20DocumentParser(root: Root, declarationParser: AsyncDeclarationParser)(
+    override implicit val ctx: AsyncWebApiContext
+) extends AsyncApiDocumentParser(root, declarationParser)(ctx)
