+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/domain/declarations/Async24DeclarationParser.scala
@@ -1,0 +1,11 @@
+package amf.apicontract.internal.spec.async.parser.domain.declarations
+
+import amf.apicontract.internal.spec.async.parser.context.AsyncWebApiContext
+import org.yaml.model.YMap
+
+object Async24DeclarationParser extends AsyncDeclarationParser {
+  override def parseDeclarations(map: YMap, parent: String)(implicit ctx: AsyncWebApiContext): Unit = {
+    Async23DeclarationParser.parseDeclarations(map, parent)
+    // TODO: add stuff....
+  }
+}
